Move `Util` into diffgraph package (its only user).
